Refactor schema layout for clarity
