+++ geometa/schema/__init__.py
@@ -1,0 +1,2 @@
+from .base import BaseSchema
+from .opensearchgeo import OpenSearchGeoSchema
